mtd-utils: fectest: Fix time formatting with _TIME_BITS=64 on 32-bit system
fectest.c formats a struct timeval with the assumption that time_t and
suseconds_t are aliases for long, but some 32-bit systems now define
them as long long in order to support timestamps beyond 2038.

As this is an elapsed time and not a timestamp, both fields should be
within the range of long, so cast them to long before formatting
rather than moving to long long.

+++ misc-utils/fectest.c
@@ -87,6 +87,6 @@ int main(void)
 		exit(1);
 	}
 
-	printf("Decoded in %ld.%06lds\n", now.tv_sec, now.tv_usec);
+	printf("Decoded in %ld.%06lds\n", (long)now.tv_sec, (long)now.tv_usec);
 	return 0;
 }
